requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
feedparser
crewai
schedule
smtplib
